requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-dotenv
google-auth
requests
